==> analysis/CNV/HRD_Aneuploidy_scores/07_compare_PR_scores.R <==
#Compare aneuploidy and HRD scores for primary and recurrent cohorts
#Required inputs:
  #txt file of aneuploidy scores in primary tumors
  #txt file of aneuploidy scores in recurrences
  #excel spreadsheet pairing up primary and recurrent tumors for comparisons
    #one recurrence chosen at random for any patients with multiple recurrences
  #HRDstats.txt file for each tumor, stored in "HRD_scores_byTumor/" directory
    #these files generated by getHRDstats_list.R
#Outputs
  #a txt file of HRD score and aneuploidy score comparisons for tumor pairs
  #statistical tests and plots for score comparisons
    #for exploration; not saved as files
  
library(readr)
library(readxl)
library(ggpubr)

my_theme<-theme(
  axis.title = element_text(size=12),
  axis.title.x = element_text(size=18),
  axis.title.y = element_text(size=18),
  axis.text.x=element_text(size=16),
  axis.text.y=element_text(size=12)
)

################################### PART 1 : Read in the data ########################################################

#Read in the tumor comparisons; start the master dataframe 
pr_comparisons<-read_xlsx("primary_rec_pairs_for_comparisons.final.xlsx",trim_ws = TRUE)

#Read in the batched aneuploidy scores, then build in columns for loading the HRD scores
primary_scores<-read_delim("Aneuploidy_scores_primaries.txt",delim="\t")
names(primary_scores)<-c("p.tumor","p.aneuploidy_score")
primary_scores$p.aneuploidy_score<-as.integer(primary_scores$p.aneuploidy_score)
primary_scores$p.HRD.Score.sum<-NA
primary_scores$p.HRD.Score.average<-NA
primary_scores$p.LST.raw<-NA
primary_scores$p.LOH.raw<-NA
primary_scores$p.NTAI.norm<-NA
primary_scores$p.NTAI.raw<-NA
primary_scores<-as.data.frame(primary_scores)

#Add in the HRD scores for each primary tumor 
for (i in 1:length(primary_scores$p.tumor))
{
  hrd_stats<-read.delim(file=paste0("HRD_scores_byTumor/",primary_scores[i,1],"_HRDstats.txt"))
  #Find the row in the primary_scores dataframe to fill in (index)
  tumor_index<-which(primary_scores$p.tumor==primary_scores[i,1])
  #Fill in HRD.Score.sum
  primary_scores[tumor_index,3]<-hrd_stats[1,1]
  #Fill in HRD.Score.average
  primary_scores[tumor_index,4]<-hrd_stats[1,2]
  #Fill in LST.raw
  primary_scores[tumor_index,5]<-hrd_stats[1,3]
  #Fill in LOH.raw
  primary_scores[tumor_index,6]<-hrd_stats[1,4]
  #Fill in NTAI.norm
  primary_scores[tumor_index,7]<-hrd_stats[1,5]
  #Fill in NTAI.raw
  primary_scores[tumor_index,8]<-hrd_stats[1,6]
}

#Repeat for recurrences dataframe 
rec_scores<-read.delim("Aneuploidy_scores_recs.txt")

names(rec_scores)=c("r.tumor","r.aneuploidy_score")
rec_scores$r.aneuploidy_score<-as.integer(rec_scores$r.aneuploidy_score)
rec_scores$r.HRD.Score.sum<-NA
rec_scores$r.HRD.Score.average<-NA
rec_scores$r.LST.raw<-NA
rec_scores$r.LOH.raw<-NA
rec_scores$r.NTAI.norm<-NA
rec_scores$r.NTAI.raw<-NA
rec_scores<-as.data.frame(rec_scores)
#Remove the rows representing multiple recurrences(need to do this to satisfy independent samples assumption of Wilcoxon test)
#see pr_comparisons for which recurrences I kept 
rec_scores<-rec_scores[-c(6,7,14,17,18,19,25,26,27,28,32,36,39),]
#Add in the HRD scores for each recurrent tumor 
for (i in 1:length(rec_scores$r.tumor))
{
  hrd_stats<-read.delim(file=paste0("HRD_scores_byTumor/",rec_scores[i,1],"_HRDstats.txt"))
  #Find the row in the rec_scores dataframe to fill in (index)
  tumor_index<-which(rec_scores$r.tumor==rec_scores[i,1])
  #Fill in HRD.Score.sum
  rec_scores[tumor_index,3]<-hrd_stats[1,1]
  #Fill in HRD.Score.average
  rec_scores[tumor_index,4]<-hrd_stats[1,2]
  #Fill in LST.raw
  rec_scores[tumor_index,5]<-hrd_stats[1,3]
  #Fill in LOH.raw
  rec_scores[tumor_index,6]<-hrd_stats[1,4]
  #Fill in NTAI.norm
  rec_scores[tumor_index,7]<-hrd_stats[1,5]
  #Fill in NTAI.raw
  rec_scores[tumor_index,8]<-hrd_stats[1,6]
}    

#Bind the dataframes together 
pr_comparisons<-cbind(pr_comparisons,primary_scores)
pr_comparisons<-cbind(pr_comparisons,rec_scores)
#visually inspect the master dataframe, then drop the duplicate tumor columns
pr_comparisons<-pr_comparisons[,-c(4,13)]
write_delim(pr_comparisons,path="PR_HRD_Aneuploidy_comparisons.txt",delim="\t",append=FALSE)
################################ PART 2 : Compare Aneuploidy scores ############################################################

wilcox.test(x=pr_comparisons$p.aneuploidy_score,y=pr_comparisons$r.aneuploidy_score,alternative="two.sided",paired=TRUE)

#Make a paired boxplot with the p-value from the wilxocon test
aneuploidy_scores<-data.frame(Primary=pr_comparisons$p.aneuploidy_score,Recurrence=pr_comparisons$r.aneuploidy_score)
aneuploidy_plot<-ggpaired(data=aneuploidy_scores,cond1="Primary",cond2="Recurrence", fill="condition",palette=c("indianred1","dodgerblue"))
aneuploidy_plot<-aneuploidy_plot+
  my_theme+
  scale_y_continuous(breaks=seq(from = 0, to = 31, by = 10),limits=c(0,31))+
  xlab("Tumor Type")+
  ylab("Total Aneuploidy Score")+
  rremove("legend")+
  geom_segment(aes(x = 1, y = 30, xend = 2, yend = 30))+
  geom_text(x=1.5,y=31,label = "ns",size=5)
aneuploidy_plot

################################ PART 3 : Compare HRD scores  ##################################################################

#Repeat Wilcoxon test and plotting for HRD score sum and all other HRD metrics 
wilcox.test(x=pr_comparisons$p.HRD.Score.sum,y=pr_comparisons$r.HRD.Score.sum,alternative="two.sided",paired=TRUE)
hrd_sum<-data.frame(Primary=pr_comparisons$p.HRD.Score.sum,Recurrence=pr_comparisons$r.HRD.Score.sum)
hrd_sum_plot<-ggpaired(data=hrd_sum,cond1="Primary",cond2="Recurrence", fill="condition",palette=c("indianred1","dodgerblue"))
hrd_sum_plot<-hrd_sum_plot+
  my_theme+
  xlab("Tumor Type")+
  ylab("Total HRD Score")+
  rremove("legend")+
  geom_segment(aes(x = 1, y = 101, xend = 2, yend = 101))+
  geom_text(x=1.5,y=103.5,label = "ns",size=5)
hrd_sum_plot

#HRD average score
wilcox.test(x=pr_comparisons$p.HRD.Score.average,y=pr_comparisons$r.HRD.Score.average,alternative="two.sided",paired=TRUE)
hrd_avg<-data.frame(Primary=pr_comparisons$p.HRD.Score.average,Recurrence=pr_comparisons$r.HRD.Score.average)
hrd_avg_plot<-ggpaired(data=hrd_avg,cond1="Primary",cond2="Recurrence", fill="condition",palette=c("indianred1","dodgerblue"))
hrd_avg_plot<-hrd_avg_plot+
  xlab("Tumor Type")+
  ylab("Average HRD Score")+
  rremove("legend")+
  geom_segment(aes(x = 1, y = 35, xend = 2, yend = 35))+
  geom_text(x=1.5,y=36,label = "ns",size=5)
hrd_avg_plot

#LST raw
wilcox.test(x=pr_comparisons$p.LST.raw,y=pr_comparisons$r.LST.raw,alternative="two.sided",paired=TRUE)
lstR<-data.frame(Primary=pr_comparisons$p.LST.raw,Recurrence=pr_comparisons$r.LST.raw)
lstR_plot<-ggpaired(data=lstR,cond1="Primary",cond2="Recurrence", fill="condition",palette=c("indianred1","dodgerblue"))
lstR_plot<-lstR_plot+
  xlab("Tumor Type")+
  ylab("Number of Large Scale State Transitions")+
  rremove("legend")+
  geom_segment(aes(x = 1, y = 51, xend = 2, yend = 51))+
  geom_text(x=1.5,y=52,label = "ns",size=5)
lstR_plot

#LOH
wilcox.test(x=pr_comparisons$p.LOH.raw,y=pr_comparisons$r.LOH.raw,alternative="two.sided",paired=TRUE)
loh<-data.frame(Primary=pr_comparisons$p.LOH.raw,Recurrence=pr_comparisons$r.LOH.raw)
loh_plot<-ggpaired(data=loh,cond1="Primary",cond2="Recurrence", fill="condition",palette=c("indianred1","dodgerblue"))
loh_plot<-loh_plot+
  xlab("Tumor Type")+
  ylab("Number of Regions of Loss of Heterozygosity")+
  rremove("legend")+
  geom_segment(aes(x = 1, y = 42, xend = 2, yend = 42))+
  geom_text(x=1.5,y=43,label = "ns",size=5)
loh_plot

#NTAI (normalized to ploidy of the tumor)
wilcox.test(x=pr_comparisons$p.NTAI.norm,y=pr_comparisons$r.NTAI.norm,alternative="two.sided",paired=TRUE)
ntai_norm<-data.frame(Primary=pr_comparisons$p.NTAI.norm,Recurrence=pr_comparisons$r.NTAI.norm)
ntai_norm_plot<-ggpaired(data=ntai_norm,cond1="Primary",cond2="Recurrence", fill="condition",palette=c("indianred1","dodgerblue"))
ntai_norm_plot<-ntai_norm_plot+
  xlab("Tumor Type")+
  ylab("Number of Regions of Non-Telomeric Allelic Imbalance \n (normalized to ploidy)")+
  rremove("legend")+
  geom_segment(aes(x = 1, y = 20, xend = 2, yend = 20))+
  geom_text(x=1.5,y=20.5,label = "ns",size=5)
ntai_norm_plot

#NTAI (raw)
wilcox.test(x=pr_comparisons$p.NTAI.raw,y=pr_comparisons$r.NTAI.raw,alternative="two.sided",paired=TRUE)
ntai_raw<-data.frame(Primary=pr_comparisons$p.NTAI.raw,Recurrence=pr_comparisons$r.NTAI.raw)
ntai_raw_plot<-ggpaired(data=ntai_raw,cond1="Primary",cond2="Recurrence", fill="condition",palette=c("indianred1","dodgerblue"))
ntai_raw_plot<-ntai_raw_plot+
  xlab("Tumor Type")+
  ylab("Number of Regions of Non-Telomeric Allelic Imbalance")+
  rremove("legend")+
  geom_segment(aes(x = 1, y = 41, xend = 2, yend = 41))+
  geom_text(x=1.5,y=42,label = "ns",size=5)
ntai_raw_plot

############### PART 4: Compare HRDtotal and aneuploidy scores but stratify by tumor type ####################################

breast<-pr_comparisons[c(1,3,4,5,8,15,20,21,23,24,25,26,27),]
ovary<-pr_comparisons[c(2,6,7,9,10,11,12,13,14,16,17,18,19,22),]

#Aneuploidy scores in breast
wilcox.test(x=breast$p.aneuploidy_score,y=breast$r.aneuploidy_score,alternative="two.sided",paired=TRUE)

aneuploidy_scores<-data.frame(Primary=breast$p.aneuploidy_score,Recurrence=breast$r.aneuploidy_score)
aneuploidy_plot<-ggpaired(data=aneuploidy_scores,cond1="Primary",cond2="Recurrence", fill="condition",palette=c("indianred1","dodgerblue"))
aneuploidy_plot<-aneuploidy_plot+
  scale_y_continuous(breaks=seq(from = 0, to = 31, by = 10),limits=c(0,31))+
  xlab("Breast Tumor Type")+
  ylab("Total Aneuploidy Score")+
  rremove("legend")+
  geom_segment(aes(x = 1, y = 30, xend = 2, yend = 30))+
  geom_text(x=1.5,y=31,label = "ns",size=5)+
  my_theme
aneuploidy_plot

#Aneuploidy scores in ovary tumors
wilcox.test(x=ovary$p.aneuploidy_score,y=ovary$r.aneuploidy_score,alternative="two.sided",paired=TRUE)

aneuploidy_scores<-data.frame(Primary=ovary$p.aneuploidy_score,Recurrence=ovary$r.aneuploidy_score)
aneuploidy_plot<-ggpaired(data=aneuploidy_scores,cond1="Primary",cond2="Recurrence", fill="condition",palette=c("indianred1","dodgerblue"))
aneuploidy_plot<-aneuploidy_plot+
  scale_y_continuous(breaks=seq(from = 0, to = 31, by = 10),limits=c(0,31))+
  xlab("Ovarian Tumor Type")+
  ylab("Total Aneuploidy Score")+
  rremove("legend")+
  geom_segment(aes(x = 1, y = 30, xend = 2, yend = 30))+
  geom_text(x=1.5,y=31,label = "ns",size=5)+
  my_theme
aneuploidy_plot

#HRDtotal in breast 
wilcox.test(x=breast$p.HRD.Score.sum,y=breast$r.HRD.Score.sum,alternative="two.sided",paired=TRUE)
hrd_sum<-data.frame(Primary=breast$p.HRD.Score.sum,Recurrence=breast$r.HRD.Score.sum)
hrd_sum_plot<-ggpaired(data=hrd_sum,cond1="Primary",cond2="Recurrence", fill="condition",palette=c("indianred1","dodgerblue"))
hrd_sum_plot<-hrd_sum_plot+
  xlab("Breast Tumor Type")+
  ylab("Total HRD Score")+
  rremove("legend")+
  geom_segment(aes(x = 1, y = 101, xend = 2, yend = 101))+
  geom_text(x=1.5,y=103.5,label = "ns",size=5)+
  my_theme
hrd_sum_plot

#HRDtotal in ovca 
wilcox.test(x=ovary$p.HRD.Score.sum,y=ovary$r.HRD.Score.sum,alternative="two.sided",paired=TRUE)
hrd_sum<-data.frame(Primary=ovary$p.HRD.Score.sum,Recurrence=ovary$r.HRD.Score.sum)
hrd_sum_plot<-ggpaired(data=hrd_sum,cond1="Primary",cond2="Recurrence", fill="condition",palette=c("indianred1","dodgerblue"))
hrd_sum_plot<-hrd_sum_plot+
  xlab("Ovarian Tumor Type")+
  ylab("Total HRD Score")+
  rremove("legend")+
  geom_segment(aes(x = 1, y = 101, xend = 2, yend = 101))+
  geom_text(x=1.5,y=103.5,label = "ns",size=5)+
  my_theme
hrd_sum_plot
